Remove unused disposedException
