+++ lib/src/redux_component/page.dart
@@ -4,7 +4,6 @@ import '../../fish_redux.dart';
 import '../redux/redux.dart';
 import 'basic.dart';
 import 'dependencies.dart';
-import 'provider.dart';
 
 /// init store's state by route-params
 typedef InitState<T, P> = T Function(P params);
@@ -159,4 +158,30 @@ class _PageState<T> extends State<_PageWidget<T>> {
     _store.teardown();
     super.dispose();
   }
+}
+
+class PageProvider extends InheritedWidget {
+  final MixedStore<Object> store;
+
+  /// Used to store page data if needed
+  final Map<String, Object> extra;
+
+  const PageProvider({
+    @required this.store,
+    @required this.extra,
+    @required Widget child,
+    Key key,
+  })  : assert(store != null),
+        assert(child != null),
+        super(key: key, child: child);
+
+  static PageProvider tryOf(BuildContext context) {
+    final PageProvider provider =
+        context.inheritFromWidgetOfExactType(PageProvider);
+    return provider;
+  }
+
+  @override
+  bool updateShouldNotify(PageProvider oldWidget) =>
+      store != oldWidget.store && extra != oldWidget.extra;
 }
